requirements.txt
PyQt6
pillow
pyinstaller